add /add-friend-request API to Controller.friend.request.AddFriendRequest
